🎨 Make header component extendable
