+++ components/Header/Header.jsx
@@ -1,0 +1,32 @@
+import Head from "next/head";
+import React from "react";
+
+const Header = ({ title, url, previewCard, description, keywordsContent }) => (
+  <Head>
+    <meta charSet="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=1,user-scalable=no"
+    />
+    <title>{title}</title>
+    <meta name="title" content={title} />
+    <meta name="description" content={description} />
+    <meta name="keywords" content={keywordsContent} />
+
+    {/* Open Graph */}
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content={url} />
+    <meta property="og:title" content={title} />
+    <meta property="og:description" content={description} />
+    <meta property="og:image" content={previewCard} />
+
+    {/* Twitter*/}
+    <meta property="twitter:card" content={previewCard} />
+    <meta property="twitter:url" content={url} />
+    <meta property="twitter:title" content={title} />
+    <meta property="twitter:description" content={description} />
+    <meta property="twitter:image" content={previewCard} />
+  </Head>
+);
+
+export default Header;
