+++ components/Header/index.js
@@ -1,0 +1,14 @@
+import Header from "./Header";
+import keywordsContent from "./keywords";
+
+const DefaultHeader = () => (
+  <Header
+    title={"Queen's COMPSA"}
+    url={"https://compsa.ca/"}
+    previewCard={"https://compsa.ca/preview.jpeg"}
+    description={"Homepage for Queen's Computing Students' Association"}
+    keywordsContent={keywordsContent}
+  />
+);
+
+export default DefaultHeader;
