+++ components/Header/keywords.js
@@ -1,0 +1,22 @@
+const keywords = [
+  "Queen's",
+  "Queen's University",
+  "Queen's University COMPSA",
+  "Queen's University Computing Society",
+  "Queen's University Computer Science Society",
+  "Queen's University Computing",
+  "Queen's University Computer Science",
+  "COMPSA",
+  "Computing Society",
+  "Computer Science Society",
+  "Computing",
+  "Computer Science",
+  "CS",
+  "CS Society",
+  "CS Club",
+  "Computer Science Club",
+];
+
+const keywordsContent = keywords.join(", ");
+
+export default keywordsContent;
